refactor: component name and folder
