+++ src/App.tsx
@@ -1,6 +1,6 @@
 
-import Header from "./header/header";
-import Main from "./main/main";
+import Header from "./components/Header/Header";
+import Main from "./components/ListView/ListView";
 
 
 export default function App() {
