+++ src/components/Pagination/Pagination.css
@@ -1,22 +1,29 @@
 .button-pagination-wrapper {
   display: flex;
+  gap: 8px;
   justify-content: center;
   padding-top: 8px;
+  flex-direction: column;
+  align-items: center;
 }
 
-.button-pagination-next, .button-pagination-back {
+.pagination-wrapper {
+  display: flex;
+}
+
+.button-pagination-next, .button-pagination-back, .button-pagination-set {
   cursor: pointer;
   border: 1px solid #000;
   background: transparent;
   height: 32px;
 }
 
-.button-pagination-next {
+.button-pagination-next, .button-pagination-set:last-of-type {
   border-radius: 0px 4px 4px 0px;
   border-left-width: 0px;
 }
 
-.button-pagination-back {
+.button-pagination-back, .button-pagination-set:first-of-type {
   border-radius: 4px 0px 0px 4px;
   border-right-width: 0px;
 }
